swig: fix fail-to-fail in %useValueEquality and %usePointerEquality
In the event of an exception from the equality operator, these would
return a class (and hence evaluate to True), rather than raise an
exception.  Might have had these raise a NotImplemented exception,
but I think that it's better to just allow any exception to propagate
up, as an exception generated by _eq_impl may not be due to the equality
operator not being implemented.