Override defaultTestLoader.suiteClass to ensure memory tests run last
The MemoryTestCase test class must always run as the final batch
of tests in any test module. By default unittest does not run the
tests in the order in which they appear in the test module so we
override the unittest.defaultTestLoader.suiteClass class with
a callable function that sorts the tests into the correct order
and returns an updated suite.