Enable flag to indicate whether temp file is expected to be created
Sometimes you want the temp file path but that file will not exist
when the with block completes. This change adds a new optional argument
to allow this. Now if a file exists when it should not, or if a file
should exist but does not, there is a RuntimeError.