Adjust test for NERSC password file
Now test for start and end of passwd file path.

This is because it seems like the file can change location.
Recently the file moved from /var/lib/sss/mc/passwd to
/var/lib/sss/mcpath/passwd. We now test for a path starting
with /var/lib and ending with passwd.

This is a more generic version of 657944960ae3e47be318739e2112d4a4b9f1b9f9