tests: add decorators to iterate over parameters
We often iterate over multiple values in tests. We have been writing this
iteration explicitly, which is annoying (for classes, this involves
subclassing; for methods, this means the values that caused a failure
aren't clear unless special efforts are made). These decorators handle
the iteration for the user, making it much easier to write and understand
tests.