Move log constants to class properties
For now, we attach these constants to the log adapter so that
it is easier to access the VERBOSE and TRACE constants.
Do not deprecated any of them since python logging does not
deprecated any (including WARN or FATAL).
Getting the constants directly from the logger ensures
self consistency for the Task user and hides the
specific logging system from the Task.