remove use of pytest-specific functionality
The teardown_method method doesn't fire when run directly, but only under
pytest, which means that a developer running the test directly isn't
testing the functionality. Adopted an alternative implementation that
doesn't rely on pytest. It's heavier than I would like (relies on the
method being tested on teardown flipping a switch), but it gets the job
done.