Remove selected unit tests for memory reporting
Unit tests testing if the memory reporting function correctly detect an
increase in memory footprint of the process are failing sporadically.
It turned out that their underlying assumption (memory consumption will
increase after defining a large list) is not always true due to the
complexities of the Python's memory management.  For that reason I
removed those tests.