Remove the test as it is confusing something in python 3.9
This tests the deprecated interface which is soon being removed.
There is some interaction with the packages in python 3.9 causing
get_caller_name to return the wrong answer when it is itself
wrapped by @deprecated. This works fine in the non-deprecated
version and on python 3.10 and 3.8 so it seems easiest to
remove the test slightly ahead of schedule.