Provide defaults for DbAuth constructor (DM-44862)
DbAuth required either the name of the file or envvar (or both)
for its constructor, but did not specify defaults. As a result
few packages had to define those to be the same in each package.
It is better to keep these values in one location and use them
as defautlts for constructor.